Dialogue  + souples

+++ Assets/Scripts/Dialogues/DialogueStarterOnSubArenaOvercrowd.cs
@@ -1,4 +1,6 @@
+using UnityEngine;
+
 public class DialogueStarterOnSubArenaOvercrowed : DialogueStarterGeneric<OnSubArenaUpdateEvent>
 {
-    protected override void OnEventTriggered(OnSubArenaUpdateEvent e)=>PlayDialogue((int)e.p_cardinalDirection);
+    protected override void OnEventTriggered(OnSubArenaUpdateEvent e)=>PlayDialogue(e.p_cardinalDirection);
 }
